+++ src/com/upyun/block/api/listener/LoadingProgressListener.java
@@ -1,5 +1,5 @@
 package com.upyun.block.api.listener;
 
 public interface LoadingProgressListener {
-	void onProgress(int bytesWritten, int totalSize);
+	void onProgress(long bytesWritten, long totalSize);
 }
